# api/swagger.yaml
openapi: 3.0.3
info:
  title: Athlete Hub API
  description: |
    REST API for the Athlete Hub sports events platform.

    ## Authentication
    - **Protected routes** require `Authorization: Bearer <session_token>`
    - **Public routes** require no authentication

    ## Base URL
    All endpoints are prefixed with `/api`.
  version: "1.0.0"
  contact:
    name: Athlete Hub Dev
servers:
  - url: /api
    description: Same-origin (development & production)
...
paths:
...
  /sites/by-subdomain/{subdomain}:
    get:
      tags: [OrganizerSites]
      summary: Public published organizer microsite by vanity subdomain
      description: >
        Returns theme, sections, legal docs, and upcoming published public events.
        Draft / pending_approval events are never included.
      parameters:
        - name: subdomain
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Site payload including events[]
        "404":
          description: Site not found or not published
  /organizer/site:
    get:
      tags: [OrganizerSites]
      summary: Load organizer microsite for the site editor
      description: >
        Authenticated organizer site chrome plus upcoming events including draft
        and pending_approval (editor layout preview only). Publish site does not
        publish events. Legal documents (terms/privacy/refund) are edited via
        PATCH with legal[] — required (ES, non-empty HTML) before event
        submit/approve.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Site payload including events[] and legal[]
        "401":
          description: Unauthorized
    patch:
      tags: [OrganizerSites]
      summary: Save organizer microsite draft or publish the site
      description: >
        Updates theme/sections/legal and optional site status (draft|published|suspended).
        Does not change event statuses — events require their own publish/approval flow
        and organizer ES legal docs (terms, privacy, refund) must be non-empty HTML.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subdomain:
                  type: string
                status:
                  type: string
                  enum: [draft, published, suspended]
                localeDefault:
                  type: string
                  enum: [es, en]
                theme:
                  type: object
                sections:
                  type: array
                  items:
                    type: object
                legal:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: Updated site payload including events[]
        "400":
          description: Invalid subdomain or payload
        "409":
          description: Subdomain taken
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...